test: splits dashboardsView into two files to add dashboardsVariables test
